==> week13/8_Project_Discussion.Rmd <==
---
title: "Project_Discussion"
author: "Rebecca Waag"
date: "11 7 2023"
output: html_document
---

# Summary & Discussion

Here, we focused exclusively on the effect of acute stress on the chromatin landscape. Analyzing ATAC-Seq data (Caradonna et al., 2022) from mice that underwent 6 minutes of cold swim stress and were euthanized 40 minutes after the end of stress, showed 911 differentially accessible regions in the nucleosome-free fraction. Interestingly, 868 of these regions showed increased accessibility, indicating more open chromatin and therefore the possibility for enhanced interaction with transcription factors. Performing enrichment analyses on regions showing increased accessibility upon stress, we find the response to glucocorticoids within the significantly enriched terms. Similarly, estimating transcription factor activity from the accessibility data, GR/GCR appears as one of the top hits. This is in line with a huge literature on GR release upon stress from the Hypothalamic–pituitary–adrenal (HPA) axis (A M Koning et al., 2019). 

Analyzing a GR/MR ChIP dataset (Mifsud et al., 2021), we find 3067 regions with stress-induced GR binding and 1642 regions with stress-induced MR binding. Most significant regions show increased TF binding upon stress exposure (GR: 1975/3067 MR: 1291/1642). 888 of these regions with increased TF binding are bound by both GR and MR, which fits with the assumption of both TFs binding to overlapping loci in the genome. 

Next, we aimed to analyze the overlap between chromatin accessibility and transcription factor bound regions upon stress. From the 868 regions with increased accessibility upon acute stress, 209 regions overlap with the significantly bound GR ChIP peaks upon stress (24% of increased accessible peaks). Of these 209 regions, 130 are also bound by MR upon stress. 79 and 8 regions are also bound by GR and MR, respectively. 651 regions with increased accessibility upon stress show no binding to GR or MR. Interestingly, these peaks also fail to reproduce the observed enrichment of glucocorticoid responsiveness using enrichment analysis (however, this might be due to the technical setup of the rGREAT server). Visualizing the coverage of control and acute stress animals at the regions that were either classified as not bound by GR or MR (651 regions) or classified as overlapping with GR peaks (209 regions), shows consistently higher signal in the FSS group in the center. However, when performing clustering, no obvious patterns that differ in the stress response to the peaks arise between the clusters. Taken together, this indicates and overall increase in binding of GR to the genome at those regions. To test whether the stress-responsive chromatin accessibility peaks show different motifs depending on whether they also bind to GR or not, we perform motif enrichment analysis. However, we find no differences in the observed motif enrichment, yielding a very similar enrichment of GCR and related motifs. Finally, we test whether known GR motifs occur at different frequencies in the stress-responsive accessibility peaks that do or do not bind GR/MR as well. While 26.3% of regions showing a signal in both the ATAC-Seq and GR-ChIP dataset contains the chosen GCR motif, only 19% of the regions showing only open chromatin do, which hints towards differences in motif occurrence between the two groups of peaks.

Several factors make the interpretation of the results difficult: 1) the ChIP dataset was obtained from rats, during the conversion to the mouse coordinates which is required for the combined analyses, several peaks were lost. 2) ChIP sequencing strongly depends on several regions, e.g. the antibody, which can lead to strong differences between individual antibodies, therefore, the comparison between GR and MR is difficult. 3) different stress models and different timepoints were used in the ATAC-Seq and ChIP-Seq experiment, this makes the comparison between those two datasets more difficult. It cannot be excluded that accessibility peaks that do not show a GR/MR signal might did to another timepoint, or did to a lesser extent, which was not captured by the antibody chosen. 

However, how the regions of GR/MR signal and chromatin accessibility act together remains an open question. Similarly, one cannot conclude from solely this data that GR is causing any potential gene regulatory effects within the open regions. The integrative analysis with available transcriptional data after a similar stressor (von Ziegler 2022) could provide more insights in this regard. However, such analyses are difficult, as chromatin accessibility is considered to precede transcription. Similarly, the question of whether the observed effects are common in all cell types or whether some are specific to certain cell types, which might provide another explanation for the disconnect between accessibility and transcription factor binding peaks, remains.

# References

A M Koning, A.-S.C., Buurstede, J.C., C M van Weert, L.T., and Meijer, O.C. (2019). Glucocorticoid and Mineralocorticoid Receptors in the Brain: A Transcriptional Perspective.

Caradonna, S.G., Paul, M.R., and Marrocco, J. (2022). An allostatic epigenetic memory on chromatin footprints after double-hit acute stress. Neurobiol. Stress 20, 100475.

Mifsud, K.R., Kennedy, C.L.M., Salatino, S., Sharma, E., Price, E.M., Haque, S.N., Gialeli, A., Goss, H.M., Panchenko, P.E., Broxholme, J., et al. (2021). Distinct regulation of hippocampal neuroplasticity and ciliary genes by corticosteroid receptors. Nat. Commun. 12.

von Ziegler, L.M., Floriou-Servou, A., Waag, R., Das Gupta, R.R., Sturman, O., Gapp, K., Maat, C.A., Kockmann, T., Lin, H.Y., Duss, S.N., et al. (2022). Multiomic profiling of the acute stress response in the mouse hippocampus. Nat. Commun. 2022 131 13, 1–20.
